Add files via upload
Devlin K-series keycap, spherical shape and non-sculpted profile. Stp file. Measurements taken with digital calipers, of product sample from "Mint green" set. 


Model profile,measurements and makeup do not belong to the author; this content is a public reference set and as such belongs to the community of hobbyists and tinkerer's of the DIY community. No data sourced from this repository is or is-to-be used for financial gain. All rights belong to original creators and copyright holders, etc. 
Information and data provided here are for reference only. Precise model and datasheet measurements are implied and may only resemble the original product by coincidence. Specifics on product-precision, material details, and engineering/manufacturing techniques belong to the original product creators/owners, unless otherwise specified.